2822: Add SpriteParser
This parser uses freeimage to load an image and creates a model that
just consists of two triangles forming a rectangle. The model renderer
renders the image on the XY plane such that the origin is in the image
center. EntityModelRenderer will orient the image towards the camera
(aka billboard rendering).