2102: Make Expression immutable
We want to make Expression immutable so that we can store the expression
by shared pointer internally. This allows us to refer to expressions
from values later on.

The only non-const member function was optimize(), which now returns the
optimized expression instead of optimizing the expression in place.

Using shared_ptr to store the expression makes it possible to use all
defaulted copy / move constructors / assignments and destructor.